Create pull-records-from-WoRMS
This code uses R to query the World Register of Marine Species based on a vector input of scientific names.  Queries are based on fuzzy names and the output is the full WoRMS record for that scientific name and all synonyms, former names, unaccepted names, and common misspellings that exist in the database.